Add the possibility to download zip archives (closes #67)
The zip archive feature is implemented in the following way:
The archive is update whenever the files in the image set are
changed (deletion or upload). The updating method observes
whether files have been added or removed and adds or removes
the respective files in the archive. This update process is
also executed when the image set is download to ensure that
the archive and the image set contain the same files.

Additionally, messages are shown to the user when the zip file
generation is still in progress (only possible after a large
image upload) or the zip file did not exist before (only possible
for sets created before this feature was introduced).

To make sure that no two threads working on the same zip file
may be executed at the same time, a database field, 'zip_finished',
is introduced, that serves this purpose.